+++ testdata/apps/full/app/setup.go
@@ -1,3 +1,5 @@
 package app
 
 func Setup() {}
+
+func Config() {}